Create Return Player Shot method, and update Runner file.
