+++ runner.rb
@@ -1,9 +1,4 @@
-require './lib/ship'
-require './lib/cell'
-require './lib/board'
-require './lib/start'
+require './lib/game.rb'
 
-start = Start.new
-start.start_game
 
-#Game.new.start
+Game.new.start
